requirements.txt
Django==4.2.4
djangorestframework==3.14.0
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
python-decouple==3.8
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
transformers==4.31.0
